Add retry queue based on dead-lettering and message TTL
It is now possible to schedule a retry if the handler fails to process the message.

The scheduling part is based on message TTL and [dead-lettering functionality of RabbitMQ](https://www.rabbitmq.com/dlx.html).
TTL for each message is calculated using following formula:

    (retry_count ** 4) + 15 + (rand(30) * (retry_count + 1))

It is important to keep in mind that the messages get expired only when they hit the head of the queue. For example, if you
have a message with 1h TTL on the head of the queue, messages after that will expire only after the one in the head is.
So the TTL times should not be very long.

The general idea is simple. For each work queue there is one additional queue used to keep messages scheduled for retry.
All messages in retry queues have specific TTL. Once a message is expired, it is automatically published back to the work queue.

In order to avoid infinite retry loops all failed messages are retried only 10 times. After ten retries
the failing message is rejected from the queue and an exception is raised.